backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langchain-qdrant
langchain-groq
pymupdf
chromadb
huggingface-hub
python-dotenv
langchain-huggingface
sentence-transformers
langchain-chroma
pydantic-settings
